update DTO, dothocbong Controler

+++ BE/DTOs/Request/ImportGpaRequest.cs
@@ -5,7 +5,7 @@ public class ImportGpaRequest
     public string MaSV { get; set; } = null!;
     public int HocKy { get; set; }
     public string NamHoc { get; set; } = null!;
-    public double GPA { get; set; }
+    public float GPA { get; set; }
     public int SoTC { get; set; }
 }
